+++ dvha/dialogs/main.py
@@ -115,6 +115,7 @@ class AddEndpointDialog(wx.Dialog):
         self.radio_box_units = wx.RadioBox(self, wx.ID_ANY, "", choices=["cc ", "% "], majorDimension=1,
                                            style=wx.RA_SPECIFY_ROWS)
         self.button_ok = wx.Button(self, wx.ID_OK, "OK")
+        self.button_ok.Disable()
         self.button_cancel = wx.Button(self, wx.ID_CANCEL, "Cancel")
 
         self.__do_bind()
@@ -155,8 +156,7 @@ class AddEndpointDialog(wx.Dialog):
         sizer_input.Add(sizer_input_units, 1, wx.ALL | wx.EXPAND, 5)
         sizer_wrapper.Add(sizer_input, 0, wx.ALL | wx.EXPAND, 10)
 
-        # TODO: Short-hand not updating? At least not on MSW?
-        self.text_short_hand = wx.StaticText(self, wx.ID_ANY, "\tShort-hand: ")
+        self.text_short_hand = wx.StaticText(self, wx.ID_ANY, "Short-hand: ")
         sizer_wrapper.Add(self.text_short_hand, 0, wx.ALL, 5)
         sizer_buttons.Add(self.button_ok, 0, wx.ALL, 5)
         sizer_buttons.Add(self.button_cancel, 0, wx.ALL | wx.EXPAND, 5)
@@ -182,7 +182,7 @@ class AddEndpointDialog(wx.Dialog):
 
     def update_label_input(self):
         new_label = "%s (%s):" % (['Input Dose', 'Input Volume']['Dose' in self.combo_box_output.GetValue()],
-                                  self.radio_box_units.GetItemLabel(self.radio_box_units.GetSelection()))
+                                  self.radio_box_units.GetItemLabel(self.radio_box_units.GetSelection()).strip())
         self.label_input_value.SetLabelText(new_label)
 
     def update_radio_box_choices(self):
@@ -190,16 +190,24 @@ class AddEndpointDialog(wx.Dialog):
         self.radio_box_units.SetItemLabel(0, choice_1)
 
     def update_short_hand(self):
-        short_hand = ['\tShort-hand: ']
-        if self.text_input.GetValue():
-            try:
-                str(float(self.text_input.GetValue()))
-                short_hand.extend([['V_', 'D_']['Dose' in self.combo_box_output.GetValue()],
-                                   self.text_input.GetValue(),
-                                   self.radio_box_units.GetItemLabel(self.radio_box_units.GetSelection()).strip()])
-            except ValueError:
-                pass
-        self.text_short_hand.SetLabelText(''.join(short_hand))
+        short_hand = 'Short-hand: '
+        value = self.text_input.GetValue()
+        if value:
+            prepend = ['V_', 'D_']['Dose' in self.combo_box_output.GetValue()]
+            units = self.radio_box_units.GetItemLabel(self.radio_box_units.GetSelection()).strip()
+            short_hand = short_hand + prepend + value + units
+        self.text_short_hand.SetLabelText(short_hand)
+        self.set_button_ok_enable(value)
+
+    def set_button_ok_enable(self, value):
+        try:
+            if value.isdigit():
+                value = int(value)
+            else:
+                value = float(value)
+        except ValueError:
+            self.text_short_hand.SetLabelText('Short-hand: ')
+        self.button_ok.Enable(type(value) in [int, float])
 
     @property
     def is_endpoint_valid(self):