+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 v0.7.4 (TBD)
 --------------------
+ - [Endpoints] Fixed a bug where endpoint short-hand would not display on MS Windows
  - [Import] Plan complexity calculation corrected, now is direct sum of all CP complexities
  - [Import] Catch PatientSetupSequence error to allow for brachytherapy plan import
  - [Import] Implement scipy.nd_image.map_coordinates for ~2.5x faster dose interpolation